HV-1551 Improve TypeHelper#isAssignable() for non parameterized types
